refactor(fsutil): fail-fast on unresolvable symlinks and share symlink tests
resolveWriteTarget now treats Lstat as the primary fact: EvalSymlinks only
runs when the caller's final path component is itself a symlink, so plain
files never pay for intermediate-directory canonicalization. A live symlink
whose resolution fails for any reason other than a missing target (a circular
link, a permission error) now fails fast at resolution time, closing the
theoretical race where EvalSymlinks fails but a later os.Stat succeeds and the
symlink is destroyed. Dangling links still return (path, true, nil) and get
replaced, preserving the existing contract.

The resolve error names the caller's path directly, so it bypasses the
"write through symlink" defer wrapper (registered later in AtomicWriteFile)
and avoids double-wrapping.

The six symlink tests move from the !windows atomicwrite_test.go into a new
build-tag-free atomicwrite_symlink_test.go gated by a requireSymlinks helper
that skips when symlink creation is unsupported. Windows CI runners lack
SeCreateSymbolicLinkPrivilege, so these tests skip there; they run on POSIX CI
and on privileged/Developer-Mode Windows machines. CircularSymlinkFailsSafely
now asserts the fail-fast "resolve symlink <path>" error instead of the
deferred "write through symlink" wrapper.
